Move tests helper classes into test folder
