+++ tests/DataLoader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Keboola\DbExtractor\Test;
+namespace Keboola\DbExtractor\Tests;
 
 use PDO;
