+++ tests/Old/CommonExtractorTest.php
@@ -8,8 +8,8 @@ use Keboola\Csv\CsvReader;
 use Keboola\DbExtractor\Application;
 use Keboola\DbExtractor\Exception\ApplicationException;
 use Keboola\Component\UserException;
-use Keboola\DbExtractor\Test\DataLoader;
-use Keboola\DbExtractor\Test\ExtractorTest;
+use Keboola\DbExtractor\Tests\DataLoader;
+use Keboola\DbExtractor\Tests\ExtractorTest;
 use Keboola\Component\Logger;
 use Monolog\Handler\TestHandler;
 use Symfony\Component\Filesystem\Filesystem;
